Убраны предупреждения компиляции.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@4182 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/parser/TAKSplitter/APEFile.cpp
@@ -296,7 +296,7 @@ HRESULT CAPEFile::Open(CBaseSplitterFile* pFile)
 				delete [] p;
 			}
 
-			if (m_frames[ape.totalframes - 1].size > tag_size) {
+			if (m_frames[ape.totalframes - 1].size > (int)tag_size) {
 				m_endpos -= tag_size;
 			}
